Dedupe viewer data source
